PPPLMER-88948: Changed the datatype from Integer to Long to handle large number values

+++ paypalhttp/src/main/java/com/paypal/http/serializer/Json.java
@@ -274,7 +274,7 @@ public class Json implements Serializer {
 		} else if (s.equals("true") || s.equals("false")) {
 			return Boolean.parseBoolean(s);
 		} else if (Character.isDigit(s.charAt(0))) {
-			return Integer.parseInt(s);
+			return Long.parseLong(s);
 		} else {
 			throw new JsonParseException("Invalid value " + s);
 		}
